requirements.txt
modoboa-admin>=1.1.1
django-braces
dnspython
